+++ qtNanomite/qtDLGHeapView.cpp
@@ -49,7 +49,7 @@ qtDLGHeapView::qtDLGHeapView(QWidget *parent, Qt::WFlags flags,int iPID)
 	_iForEntry = 0;
 	_iForEnd = myMainWindow->coreDebugger->PIDs.size();
 
-	for(int i = 0; i < myMainWindow->coreDebugger->PIDs.size(); i++)
+	for(size_t i = 0; i < myMainWindow->coreDebugger->PIDs.size(); i++)
 	{
 		if(myMainWindow->coreDebugger->PIDs[i].dwPID == _iPID)
 			_iForEntry = i; _iForEnd = i +1;
@@ -215,7 +215,7 @@ void qtDLGHeapView::DisplayHeap()
 	tblHeapView->setRowCount(0);
 	tblHeapBlocks->setRowCount(0);
 
-	for(int i = _iForEntry; i < _iForEnd;i++)
+	for(size_t i = _iForEntry; i < _iForEnd;i++)
 	{
 		HEAPLIST32 heapList;
 		heapList.dwSize = sizeof(HEAPLIST32);
